fixes for buildable project

+++ src/items/itemweapon.h
@@ -4,6 +4,8 @@
 #include "item.h"
 #include <QList>
 
+#include <commonmacros.h>
+
 
 class ItemWeapon : public Item
 {
@@ -15,7 +17,7 @@ public:
     ItemWeapon(ItemWeaponAttribute parameters);
     ~ItemWeapon() override;
 
-    int getHands() override { return _AttributeWeapon.Hands;}
+    int getHands() override { return INT(_AttributeWeapon.Hands);}
     ItemWeaponDamage getDamage() override;              // Please be sure that you need to set Hand in the inventory
 
 };
